ChatApp/requirements.txt
django
django-rest-framework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
python-dotenv
channels
django-cors-headers
daphne
asgiref